test: split GuestAuthenticationFlowTest by responsibility (#37)
